requirements.txt
nltk>=3.9.1
pandas>=2.2.3
plotly>=6.0.1
pypdf2>=3.0.1
python-docx>=1.1.2
scikit-learn>=1.6.1
spacy>=3.8.5
streamlit>=1.44.1